requirements.txt
pystray==0.19.5
Pillow==10.0.1
pygame==2.5.2